requirements.txt
tortoise-orm
uvicorn
fastapi
fastapi-mail
python-dotenv